Extract resource management and new action group into standalone files
