+++ src/codemagic/tools/_app_store_connect/action_groups/__init__.py
@@ -1,0 +1,1 @@
+from .app_store_action_group import AppStoreActionGroup
