+++ src/codemagic/tools/_app_store_connect/errors.py
@@ -1,0 +1,5 @@
+from codemagic import cli
+
+
+class AppStoreConnectError(cli.CliAppException):
+    pass
